Implementation of GET request for a single user
Addition of GET request implementation for a SingleUser. This is used to
fetch a single user and return a response based on the user id provided.
If the user id is not for the current logged in user, the api shold
abort the request as the user is not Authorized to access that
information.

If the user does exist, then the UserAccount and UserProfile Model
respresentations should be returned in the response with a code of 200.

Addition of the `output_json` function to handle responses is to
properly serialize the datetime objects from these model objects as the
FlaskRESTful api does not support if from the JSON module in the Python
Standard library